Adapting admin classes for compatibility with WordPress 4.4
With the release of WordPress 4.4, wp-admin changed some of the classes used for ui elements. Adding a class of `hndle` to the H3 provides the appropriate left and right padding while decreasing the top and bottom spacing. 

Adding `<span class="toggle-indicator" aria-hidden="true"></span>` renders the arrow to click on to expand the collapsible ui element.